requirements.txt
amqp==2.5.2
asgiref==3.2.3
astroid==2.3.3
bcrypt==3.1.7
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==3.0.2
django-cors-headers==3.2.1
django-extensions==2.2.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.8
isort==4.3.21
kombu==4.6.7
lazy-object-proxy==1.4.3
mccabe==0.6.1
paramiko==2.7.1
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
PyNaCl==1.3.0
pysftp==0.2.9
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
vine==1.3.0
wrapt==1.11.2
XlsxWriter==1.2.7
xmltodict==0.12.0
